Fixed some warnings discovered by CLion.

+++ src/oofemlib/floatmatrix.C
@@ -164,11 +164,11 @@ FloatMatrix &FloatMatrix :: operator = ( std :: initializer_list< std :: initial
 
 FloatMatrix &FloatMatrix :: operator = ( std :: initializer_list< FloatArray >mat )
 {
-    RESIZE( ( int ) mat.begin()->giveSize(), ( int ) mat.size() );
+    RESIZE( mat.begin()->giveSize(), ( int ) mat.size() );
     auto p = this->values.begin();
     for ( auto col : mat ) {
 #if DEBUG
-        if ( this->nRows != ( int ) col.giveSize() ) {
+        if ( this->nRows != col.giveSize() ) {
                 OOFEM_ERROR("Initializer list has inconsistent column sizes.");
         }
 #endif
